@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 	</PropertyGroup>
 	<ItemGroup>
 	  <None Remove="Validators\" />
@@ -15,6 +15,6 @@
 	  <Folder Include="Models\" />
 	</ItemGroup>
 	<ItemGroup>
-	  <PackageReference Include="libphonenumber-csharp" Version="8.12.38" />
+	  <PackageReference Include="libphonenumber-csharp" Version="8.13.3" />
 	</ItemGroup>
 </Project>
