@@ -5,27 +5,26 @@ using Application.Common.Interfaces;
 using MediatR.Pipeline;
 using Microsoft.Extensions.Logging;
 
-namespace Application.Common.Behaviours
+namespace Application.Common.Behaviours;
+
+public class RequestLogger<TRequest> : IRequestPreProcessor<TRequest>
 {
-    public class RequestLogger<TRequest> : IRequestPreProcessor<TRequest>
+    private readonly ILogger _logger;
+    private readonly ICurrentUserService _currentUserService;
+
+    public RequestLogger(ILogger<TRequest> logger, ICurrentUserService currentUserService)
     {
-        private readonly ILogger _logger;
-        private readonly ICurrentUserService _currentUserService;
+        _logger = logger;
+        _currentUserService = currentUserService;
+    }
 
-        public RequestLogger(ILogger<TRequest> logger, ICurrentUserService currentUserService)
-        {
-            _logger = logger;
-            _currentUserService = currentUserService;
-        }
+    public Task Process(TRequest request, CancellationToken cancellationToken)
+    {
+        var name = typeof(TRequest).Name;
 
-        public Task Process(TRequest request, CancellationToken cancellationToken)
-        {
-            var name = typeof(TRequest).Name;
+        _logger.LogInformation("Request: {@Name} {@UserId} {@Request}",
+            name, _currentUserService.UserId, request);
 
-            _logger.LogInformation("Request: {@Name} {@UserId} {@Request}",
-                name, _currentUserService.UserId, request);
-
-            return Task.CompletedTask;
-        }
+        return Task.CompletedTask;
     }
 }
