@@ -5,39 +5,38 @@ using Application.Common.Interfaces;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
-namespace Application.Common.Behaviours
+namespace Application.Common.Behaviours;
+
+public class RequestPerformanceBehaviour<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse> where TRequest : IRequest<TResponse>
 {
-    public class RequestPerformanceBehaviour<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse> where TRequest : IRequest<TResponse>
+    private readonly Stopwatch _timer;
+    private readonly ILogger<TRequest> _logger;
+    private readonly ICurrentUserService _currentUserService;
+
+    public RequestPerformanceBehaviour(ILogger<TRequest> logger, ICurrentUserService currentUserService)
     {
-        private readonly Stopwatch _timer;
-        private readonly ILogger<TRequest> _logger;
-        private readonly ICurrentUserService _currentUserService;
+        _timer = new Stopwatch();
 
-        public RequestPerformanceBehaviour(ILogger<TRequest> logger, ICurrentUserService currentUserService)
+        _logger = logger;
+        _currentUserService = currentUserService;
+    }
+
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
+    {
+        _timer.Start();
+
+        var response = await next();
+
+        _timer.Stop();
+
+        if (_timer.ElapsedMilliseconds > 500)
         {
-            _timer = new Stopwatch();
+            var name = typeof(TRequest).Name;
 
-            _logger = logger;
-            _currentUserService = currentUserService;
+            _logger.LogWarning("Long Running Request: {Name} ({ElapsedMilliseconds} milliseconds) {@UserId} {@Request}",
+                name, _timer.ElapsedMilliseconds, _currentUserService.UserId, request);
         }
 
-        public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
-        {
-            _timer.Start();
-
-            var response = await next();
-
-            _timer.Stop();
-
-            if (_timer.ElapsedMilliseconds > 500)
-            {
-                var name = typeof(TRequest).Name;
-
-                _logger.LogWarning("Long Running Request: {Name} ({ElapsedMilliseconds} milliseconds) {@UserId} {@Request}",
-                    name, _timer.ElapsedMilliseconds, _currentUserService.UserId, request);
-            }
-
-            return response;
-        }
+        return response;
     }
 }
