@@ -10,22 +10,21 @@ using MediatR.Pipeline;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Application
+namespace Application;
+
+public static class DependencyInjection
 {
-    public static class DependencyInjection
+    public static IServiceCollection AddApplication(this IServiceCollection services)
     {
-        public static IServiceCollection AddApplication(this IServiceCollection services)
-        {
-            services.AddAutoMapper(Assembly.GetExecutingAssembly());
-            services.AddMediatR(new Assembly[] { Assembly.GetExecutingAssembly(), typeof(Entity).Assembly });
+        services.AddAutoMapper(Assembly.GetExecutingAssembly());
+        services.AddMediatR(new Assembly[] { Assembly.GetExecutingAssembly(), typeof(Entity).Assembly });
 
-            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(RequestPerformanceBehaviour<,>));
-            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(RequestValidationBehavior<,>));
-            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(TransactionBehaviour<,>));
-            services.AddTransient(typeof(IRequestPreProcessor<>), typeof(RequestLogger<>));
+        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(RequestPerformanceBehaviour<,>));
+        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(RequestValidationBehavior<,>));
+        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(TransactionBehaviour<,>));
+        services.AddTransient(typeof(IRequestPreProcessor<>), typeof(RequestLogger<>));
 
-            services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-            return services;
-        }
+        services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
+        return services;
     }
 }
