@@ -1,19 +1,28 @@
 ﻿using System.Security.Claims;
 using Application.Common.Interfaces;
+using Application.Common.Models;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 
-namespace WebAPI.Services
+namespace WebAPI.Services;
+
+public class CurrentUserService : ICurrentUserService
 {
-    public class CurrentUserService : ICurrentUserService
+    public CurrentUserService(IHttpContextAccessor httpContextAccessor, IUserManager userManager)
     {
-        public CurrentUserService(IHttpContextAccessor httpContextAccessor)
+        UserId = httpContextAccessor.HttpContext?.User?.FindFirstValue(ClaimTypes.NameIdentifier);
+        IsAuthenticated = UserId != null;
+
+        if (IsAuthenticated && User is null)
         {
-            UserId = httpContextAccessor.HttpContext?.User?.FindFirstValue(ClaimTypes.NameIdentifier);
-            IsAuthenticated = UserId != null;
+            var res = userManager.FindByIdAsync(UserId).Result;
+            User = res;
         }
-
-        public string UserId { get; }
-
-        public bool IsAuthenticated { get; }
     }
+
+    public string UserId { get; }
+
+    public bool IsAuthenticated { get; }
+
+    public UserDTO User { get; }
 }
