@@ -1,133 +1,30 @@
-﻿@inject Microsoft.AspNetCore.Mvc.Localization.IViewLocalizer Localizer
-@model Application.CQRS.Rooms.Commands.Create.CreateRoomCommand
-@{
-    ViewData["Title"] = Localizer["Create Room"];
-    ViewData["Description"] = "";
-    ViewData["Keywords"] = "";
-}
+﻿@model Application.CQRS.Rooms.Commands.Create.CreateRoomCommand
+
+@inject Application.Common.Models.ConfigModels.TenantInfo tenant
+@inject Application.Common.Interfaces.ICurrentUserService CurrentUserService
+@inject Microsoft.AspNetCore.Mvc.Localization.IViewLocalizer Localizer
 
 @section Styles{
-    <link href="assets/css/pages/wizard/wizard-3.css" rel="stylesheet" type="text/css" />
-    <link href="assets/plugins/custom/uppy/uppy.bundle.css" rel="stylesheet" type="text/css" />
-
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.8/css/intlTelInput.css" />
     <script src="https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.8/js/intlTelInput.min.js"></script>
+
     <style>
         .iti {
-            display: block;
+            width: 100%;
+            @*background-color: transparent;*@
         }
 
         .iti--separate-dial-code .iti__selected-flag {
-            border-top-left-radius: 0.85rem;
-            border-bottom-left-radius: 0.85rem;
             background-color: transparent;
-        }
-
-        iti__country-list {
-            position: initial;
         }
     </style>
 }
 
 @section Scripts{
-    <script src="~/assets/js/jquery-validation/dist/jquery.validate.min.js"></script>
-    <script src="~/assets/js/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>
+    <partial name="_ValidationScriptsPartial" />
 
-
-    <script src="assets/js/pages/rooms/create.js" defer></script>
-
-    <script src="assets/plugins/custom/uppy/uppy.bundle.js"></script>
-    <script>var mediaUrls = [];
-        const Tus = Uppy.XHRUpload;
-        const ProgressBar = Uppy.ProgressBar;
-        const StatusBar = Uppy.StatusBar;
-        const FileInput = Uppy.FileInput;
-        const Informer = Uppy.Informer;
-
-        var id = '#kt_uppy_3';
-
-        var uppyDrag = Uppy.Core({
-            autoProceed: true,
-            restrictions: {
-                maxFileSize: 2000000, // 1mb
-                maxNumberOfFiles: 10,
-                minNumberOfFiles: 1,
-                allowedFileTypes: ['image/*', 'video/*']
-            }
-        });
-
-        uppyDrag.use(Uppy.DragDrop, { target: id + ' .uppy-drag' });
-        uppyDrag.use(ProgressBar, {
-            target: id + ' .uppy-progress',
-            hideUploadButton: false,
-            hideAfterFinish: false
-        });
-
-        uppyDrag.use(Informer, { target: id + ' .uppy-informer' });
-        uppyDrag.use(Tus, {
-            endpoint: '/files/upload',
-        });
-
-        uppyDrag.on('complete', function (file) {
-            var imagePreview = "";
-            $.each(file.successful, function (index, value) {
-                var imageType = /image/;
-                var thumbnail = "";
-                if (imageType.test(value.type)) {
-                    thumbnail = '<div class="uppy-thumbnail"><img src="' + value.uploadURL + '"/></div>';
-                    mediaUrls.push(value.uploadURL);
-                }
-                var sizeLabel = "bytes";
-                var filesize = value.size;
-                if (filesize > 1024) {
-                    filesize = filesize / 1024;
-                    sizeLabel = "kb";
-                    if (filesize > 1024) {
-                        filesize = filesize / 1024;
-                        sizeLabel = "MB";
-                    }
-                }
-                imagePreview += '<div class="uppy-thumbnail-container" data-id="' + value.id + '">' + thumbnail + ' <span class="uppy-thumbnail-label">' + value.name + ' (' + Math.round(filesize, 2) + ' ' + sizeLabel + ')</span><span data-id="' + value.id + '" class="uppy-remove-thumbnail"><i class="flaticon2-cancel-music"></i></span></div>';
-            });
-
-            $(id + ' .uppy-thumbnails').append(imagePreview);
-            generateMediaUrlInputs();
-        });
-
-        $(document).on('click', id + ' .uppy-thumbnails .uppy-remove-thumbnail', function () {
-            var imageId = $(this).attr('data-id');
-            var container = $(id + ' .uppy-thumbnail-container[data-id="' + imageId + '"');
-            var imgUrl = '';
-            var img = container.find("img"), // select images inside .container
-                len = img.length; // check if they exist
-            if (len > 0) {
-                // images found, get id
-                imgUrl = img.first().attr("src"); // get id of first image
-            } else {
-                // images not found
-            }
-
-            uppyDrag.removeFile(imageId);
-            container.remove();
-            mediaUrls = mediaUrls.filter(function (ele) {
-                return ele != imgUrl;
-            });
-            generateMediaUrlInputs();
-        });
-
-        function generateMediaUrlInputs() {
-            var mediaurlinputsStr = "";
-            console.log(mediaUrls);
-            mediaUrls.forEach(function (elem) {
-                mediaurlinputsStr += `<input type="checkbox" hidden name="MediaUrls" value="${elem}" checked="checked" />`;
-            });
-            $('#medias').html(mediaurlinputsStr);
-        }</script>
-
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyD1uU_tT63977_KyXVILofNxnKlLdA32go&libraries=places&callback=initMap&v=weekly" defer></script>
-    <script src="assets/js/pages/rooms/googlemap.js"></script>
-
-    <script>const phoneInputField = document.querySelector("#phone");
+    <script>
+        const phoneInputField = document.querySelector("#phone");
         const phoneInput = window.intlTelInput(phoneInputField, {
             separateDialCode: true,
             allowDropdown: true,
@@ -140,174 +37,396 @@
                 $('#ContactPhone').val(phoneInput.getNumber());
                 document.querySelector('[data-valmsg-for="ContactPhone"]').innerHTML = "";
             } else {
-                document.querySelector('[data-valmsg-for="ContactPhone"]').innerHTML = "Enter valid phone";
+                document.querySelector('[data-valmsg-for="ContactPhone"]').innerHTML = "@Localizer["Enter valid phone"]";
+            }
+        });
+    </script>
+    @if (Model is not null)
+    {
+        <script>
+            phoneInput.setNumber("@Html.Raw(Model.ContactPhone)");
+        </script>
+    }
+    <script>
+        var selectedCategoryId = '@Model?.CategoryId';
+        var selectedAmenitieIds = @Html.Raw(Json.Serialize(Model?.AmenitieIds));
+        selectedAmenitieIds = selectedAmenitieIds ? selectedAmenitieIds : [];
+        var selectedRequirementIds = @Html.Raw(Json.Serialize(Model?.RequirementIds));
+        selectedRequirementIds = selectedRequirementIds ? selectedRequirementIds : [];
+
+        const generateMegaOption = (imgUrl, title, id, isChecked, type, name) => {
+            var checked = isChecked ? 'checked' : '';
+            return `<label class="option">
+                        <span class="option-label">
+                            <span class="option-body">
+                                <img src="${imgUrl}" class="h-50px w-50px" />
+                                <span class="text-dark">${title}</span>
+                            </span>
+                        </span>
+
+                        <span class="option-control">
+                            <span class="${type}">
+                                <input type="${type}" name="${name}" value="${id}" ${checked} />
+                                <span></span>
+                            </span>
+                        </span>
+                    </label>`;
+        };
+        $.ajax({
+            url: "/rooms/categories",
+            type: 'GET',
+            success: function (res) {
+                res.forEach(function (data, index, arr) {
+                    var newOption = generateMegaOption(data.icon, data.name, data.id, selectedCategoryId == data.id, 'radio', 'CategoryId');
+                    $("#categoriesContainer").append(`<div class="col-lg-6">${newOption}</div>`);
+                });
+            }
+        });
+        $.ajax({
+            url: "/rooms/amenities",
+            type: 'GET',
+            success: function (res) {
+                res.forEach(function (data, index, arr) {
+                    var newOption = generateMegaOption(data.icon, data.name, data.id, selectedAmenitieIds.includes(data.id), 'checkbox', 'AmenitieIds');
+                    $("#amenitiesContainer").append(`<div class="col-lg-4">${newOption}</div>`);
+                });
+            }
+        });
+        $.ajax({
+            url: "/rooms/requirements",
+            type: 'GET',
+            success: function (res) {
+                res.forEach(function (data, index, arr) {
+                    var newOption = generateMegaOption(data.icon, data.name, data.id, selectedRequirementIds.includes(data.id), 'checkbox', 'RequirementIds');
+                    $("#requirementsContainer").append(`<div class="col-lg-4">${newOption}</div>`);
+                });
             }
         });</script>
+
+    <script>const customWizard = $('[data-custom-wizard]')[0];
+        var steps = $(customWizard).children('[data-custom-wizard-step]');
+        var nextBtn = $('[data-custom-wizard-next]')[0];
+        var previousBtn = $('[data-custom-wizard-previous]')[0];
+        var submitBtn = $('[data-custom-wizard-submit]')[0];
+
+        let stageCount = steps.length - 1;
+        let currentStage = 0;
+
+        const changeStage = () => {
+            $(steps).hide();
+            steps.eq(currentStage).show();
+        }
+
+        $(nextBtn).on('click', function () {
+            if (currentStage < stageCount) {
+                currentStage++;
+                changeStage();
+                resetButtons();
+            }
+        });
+        $(previousBtn).on('click', function () {
+            if (currentStage > 0) {
+                currentStage--;
+                changeStage();
+                resetButtons();
+            }
+        });
+        const resetButtons = () => {
+            $(nextBtn).hide();
+            $(submitBtn).hide();
+            $(previousBtn).hide();
+
+            if (currentStage == stageCount) {
+                $(submitBtn).show();
+            }
+            if (currentStage > 0) {
+                $(previousBtn).show();
+            }
+            if (currentStage < stageCount) {
+                $(nextBtn).show();
+            }
+        };
+        const initCustomWizard = () => {
+            resetButtons();
+            changeStage();
+        };
+        initCustomWizard();</script>
+
+    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyD1uU_tT63977_KyXVILofNxnKlLdA32go&libraries=places&callback=initMap&v=weekly"></script>
+
+    <script>
+        var geocoder = new google.maps.Geocoder();
+        var getLocationFromGPSBtn = $('#getlocationfromgpsbtn');
+        var lngInput = $('#Lng');
+        var latInput = $('#Lat');
+        var addressInput = $('#Address');
+
+        function getAddressStringFromLatLng(loc) {
+            var latlng = new google.maps.LatLng(loc.lat, loc.lng);
+            geocoder.geocode({
+                'latLng': latlng
+            }, function (results, status) {
+                if (status === google.maps.GeocoderStatus.OK) {
+                    if (results[1]) {
+                        addressInput.val(results[1].formatted_address);
+                    } else {
+                        alert('No results found');
+                    }
+                } else {
+                }
+            });
+        }
+        const updateCoordinates = (loc) => {
+            latInput.val(loc.lat);
+            lngInput.val(loc.lng);
+        }
+
+        function initSearchBox() {
+            var input = document.getElementById("Address");
+            // Create the search box and link it to the UI element
+            const searchBox = new google.maps.places.SearchBox(input);
+
+            // Listen for the event fired when the user selects a prediction and retrieve
+            // more details for that place.
+            searchBox.addListener("places_changed", () => {
+                const places = searchBox.getPlaces();
+                if (places.length == 0) {
+                    return;
+                }
+
+                places.forEach((place) => {
+                    if (!place.geometry || !place.geometry.location) {
+                        // console.log("Returned place contains no geometry");
+                        return;
+                    }
+
+                    updateCoordinates(place.geometry.location);
+                });
+            });
+        }
+        initSearchBox();
+
+        function getUserLocationFromGPS() {
+            var btnhtml = getLocationFromGPSBtn.html();
+            getLocationFromGPSBtn.html("<span class=\"spinner-border spinner-border-sm align-middle ms-2\"></span>");
+            navigator.geolocation.getCurrentPosition(
+                (position) => {
+                    pos = {
+                        lat: position.coords.latitude,
+                        lng: position.coords.longitude,
+                    };
+                    updateCoordinates(pos);
+                    getAddressStringFromLatLng(pos);
+                    getLocationFromGPSBtn.html(btnhtml);
+                },
+                (err) => {
+                    getLocationFromGPSBtn.html(btnhtml);
+                }
+            );
+        }
+
+        let askedForUsingGPSLocation = false;
+        $('#Address').on('focus', function () {
+            if (!askedForUsingGPSLocation) {
+                 Swal.fire({
+                        title: "@Localizer["Do you want to use your current location?"]",
+                        text: "@Localizer["We keep your location secret. It will not be shown to anyone."]",
+                        icon: "question",
+                     buttonsStyling: false,
+                     reverseButtons: true,
+
+                     confirmButtonText: "<i class='la la-thumbs-up'></i> @Localizer["Yes"]",
+                     cancelButtonText: "<i class='la la-pen-fancy'></i> @Localizer["No, I want to write"]",
+                        showCancelButton: true,
+                        customClass: {
+                            confirmButton: "btn fw-bold btn-success",
+                            cancelButton: "btn fw-bold btn-info",
+                        }
+                    }).then((res) => {
+                        if (res.isConfirmed) {
+                            getUserLocationFromGPS();
+                        }
+                    });
+                askedForUsingGPSLocation = true;
+            }
+        });
+        getLocationFromGPSBtn.on('click', function () {
+            getUserLocationFromGPS();
+        });
+    </script>
+
+    <script>$('#Price').TouchSpin({
+            buttondown_class: 'btn btn-secondary',
+            buttonup_class: 'btn btn-secondary',
+
+            min: 0,
+            step: 10,
+            decimals: 0,
+            boostat: 8,
+            maxboostedstep: 10,
+        });
+
+        var filePreview = $('.filePreview').clone();
+        $('.filePreview').remove();
+
+        $('#mediafileinput').on('change', function () {
+            var files = $('#mediafileinput').prop('files');
+
+            KTApp.blockPage({
+                overlayColor: '#000000',
+                state: 'primary',
+                message: '@Localizer["Uploading files. Please wait."]'
+            });
+            for (var i = 0; i != files.length; i++) {
+                let formData = new FormData();
+                formData.append("files", files[i]);
+
+                $.ajax({
+                    url: "/files/upload",
+                    data: formData,
+                    async: false,
+                    processData: false,
+                    contentType: false,
+                    type: "POST",
+                    success: function (res) {
+                        var item = filePreview.clone();
+                        item.find('img').attr('src', res.url);
+                        item.find('input').attr('value', res.url);
+                        $('#mediasContainer').append(item);
+                    }
+                });
+            }
+            KTApp.unblockPage();
+        });
+
+        $('[data-delete-uploaded-file]').on('click', function () {
+            $(this).parents('.filePreview').parent().remove();
+        });</script>
 }
+<div class="d-flex flex-column-fluid align-items-center mb-12 mb-lg-0">
+    <div class="container pt-15">
 
-
-<!--begin: Wizard-->
-<div class="wizard wizard-3" id="kt_wizard_v3" data-wizard-state="step-first" data-wizard-clickable="true">
-    <!--begin: Wizard Nav-->
-    <div class="wizard-nav">
-        <div class="wizard-steps px-8 py-8 px-lg-15 py-lg-3">
-            <!--begin::Wizard Step 1 Nav-->
-            <div class="wizard-step" data-wizard-type="step" data-wizard-state="current">
-            </div>
-            <!--end::Wizard Step 1 Nav-->
-            <!--begin::Wizard Step 2 Nav-->
-            <div class="wizard-step" data-wizard-type="step" data-wizard-state="current">
-            </div>
-            <!--end::Wizard Step 2 Nav-->
-            <!--begin::Wizard Step 3 Nav-->
-            <div class="wizard-step" data-wizard-type="step">
-            </div>
-            <!--end::Wizard Step 3 Nav-->
-            <!--begin::Wizard Step 4 Nav-->
-            <div class="wizard-step" data-wizard-type="step">
-            </div>
-            <!--end::Wizard Step 4 Nav-->
-            <!--begin::Wizard Step 5 Nav-->
-            <div class="wizard-step" data-wizard-type="step">
-            </div>
-            <!--end::Wizard Step 5 Nav-->
-        </div>
-    </div>
-    <!--end: Wizard Nav-->
-    <!--begin: Wizard Body-->
-    <div class="row justify-content-center py-10 px-8 py-lg-12 px-lg-10">
-        <div class="col-xl-12 col-xxl-7">
-            <!--begin: Wizard Form-->
-            <form class="form" id="kt_form" method="post">
-                <!--begin: Wizard Step 1-->
-                <div class="pb-5" data-wizard-type="step-content" data-wizard-state="current">
-                    <h4 class="mb-10 font-weight-bold text-dark">Enter Contact Details</h4>
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="ContactName"></label>
-                        <input asp-for="ContactName" class="form-control" />
-                        <span class="form-text text-muted" asp-validation-for="ContactName"></span>
-                    </div>
-                    <!--end::Input-->
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="ContactPhone"></label>
-                        <input id="phone" class="form-control" />
-                        <input hidden asp-for="ContactPhone" />
-                        <span class="form-text text-muted" asp-validation-for="ContactPhone"></span>
-                    </div>
-                    <!--end::Input-->
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="ContactEmail"></label>
-                        <input asp-for="ContactEmail" class="form-control" />
-                        <span class="form-text text-muted" asp-validation-for="ContactEmail"></span>
-                    </div>
-                    <!--end::Input-->
-
+        <div class="row justify-content-center">
+            <div class="col-xl-12 col-xxl-7">
+                <div class="w-100 d-flex justify-content-start mb-5">
+                    <h1 class="font-weight-bolder">@Localizer["Create announcement"]</h1>
                 </div>
-                <!--end: Wizard Step 1-->
-                <!--begin: Wizard Step 2-->
-                <div class="pb-5" data-wizard-type="step-content">
-                    <h4 class="mb-10 font-weight-bold text-dark">Setup Your Current Location</h4>
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <div class="input-group">
-                            <input id="addressinput" asp-for="Address" class="form-control" aria-label="Search address" />
-                            <div class="input-group-append">
-                                <button type="button" class="btn btn-outline-secondary" onclick="getUserLocationFromGPS()">
-                                    <span class="svg-icon svg-icon-primary">
-                                        <!--begin::Svg Icon | path:/var/www/preview.keenthemes.com/metronic/releases/2021-05-14-112058/theme/html/demo5/dist/../src/media/svg/icons/Map/Position.svg--><svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px" height="24px" viewBox="0 0 24 24" version="1.1">
-                                            <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-                                                <rect x="0" y="0" width="24" height="24" />
-                                                <path d="M19,11 L21,11 C21.5522847,11 22,11.4477153 22,12 C22,12.5522847 21.5522847,13 21,13 L19,13 C18.4477153,13 18,12.5522847 18,12 C18,11.4477153 18.4477153,11 19,11 Z M3,11 L5,11 C5.55228475,11 6,11.4477153 6,12 C6,12.5522847 5.55228475,13 5,13 L3,13 C2.44771525,13 2,12.5522847 2,12 C2,11.4477153 2.44771525,11 3,11 Z M12,2 C12.5522847,2 13,2.44771525 13,3 L13,5 C13,5.55228475 12.5522847,6 12,6 C11.4477153,6 11,5.55228475 11,5 L11,3 C11,2.44771525 11.4477153,2 12,2 Z M12,18 C12.5522847,18 13,18.4477153 13,19 L13,21 C13,21.5522847 12.5522847,22 12,22 C11.4477153,22 11,21.5522847 11,21 L11,19 C11,18.4477153 11.4477153,18 12,18 Z" fill="#000000" fill-rule="nonzero" opacity="0.3" />
-                                                <circle fill="#000000" opacity="0.3" cx="12" cy="12" r="2" />
-                                                <path d="M12,17 C14.7614237,17 17,14.7614237 17,12 C17,9.23857625 14.7614237,7 12,7 C9.23857625,7 7,9.23857625 7,12 C7,14.7614237 9.23857625,17 12,17 Z M12,19 C8.13400675,19 5,15.8659932 5,12 C5,8.13400675 8.13400675,5 12,5 C15.8659932,5 19,8.13400675 19,12 C19,15.8659932 15.8659932,19 12,19 Z" fill="#000000" fill-rule="nonzero" />
-                                            </g>
-                                        </svg><!--end::Svg Icon-->
-                                    </span>
-                                </button>
+                <form method="post">
+                    <div asp-validation-summary="All"></div>
+                    <div data-custom-wizard>
+                        <div data-custom-wizard-step style="display:none;">
+                            <h3>@Localizer["Select announcement category"]</h3>
+                            <div class="row h-100" id="categoriesContainer">
                             </div>
                         </div>
-                        <span class="form-text text-muted" asp-validation-for="Address"></span>
-                    </div>
-                    <!--end::Input-->
-                    <div id="map" style="height:500px"></div>
-                </div>
-                <!--end: Wizard Step 2-->
-                <!--begin: Wizard Step 3-->
-                <div class="pb-5" data-wizard-type="step-content">
-                    <h4 class="mb-10 font-weight-bold text-dark">Enter the Details of room</h4>
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="Name"></label>
-                        <input asp-for="Name" class="form-control" />
-                        <span class="form-text text-muted" asp-validation-for="Name"></span>
-                    </div>
-                    <!--end::Input-->
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="Description"></label>
-                        <textarea asp-for="Description" class="form-control"></textarea>
-                        <span class="form-text text-muted" asp-validation-for="Description"></span>
-                    </div>
-                    <!--end::Input-->
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="Price"></label>
-                        <input asp-for="Price" class="form-control" />
-                        <span class="form-text text-muted" asp-validation-for="Price"></span>
-                    </div>
-                    <!--end::Input-->
-                    <!--begin::Input-->
-                    <div class="form-group">
-                        <label asp-for="CategoryId"></label><br />
-                        <select asp-for="CategoryId" class="form-control select2"></select>
-                        <span class="form-text text-muted" asp-validation-for="CategoryId"></span>
-                    </div>
-                    <!--end::Input-->
-                </div>
-                <!--end: Wizard Step 3-->
-                <!--begin: Wizard Step 4-->
-                <div class="pb-5" data-wizard-type="step-content">
-                    <h4 class="mb-10 font-weight-bold text-dark">Select your Services</h4>
-                    <div class="row" id="Amenities"></div>
-                </div>
-                <!--end: Wizard Step 4-->
-                <!--begin: Wizard Step 5-->
-                <div class="pb-5" data-wizard-type="step-content">
-                    <h4 class="mb-10 font-weight-bold text-dark">Setup Your Delivery Location</h4>
-                    <div class="my-5">
-                        <div class="row" id="Requirements">
+                        <div data-custom-wizard-step style="display:none;">
+                            <h3>@Localizer["Enter location"]</h3>
+                            <div class="form-group">
+                                <div class="input-group">
+                                    <input type="text" class="form-control" asp-for="Address" placeholder="@Localizer["Enter address"]" />
+                                    <div class="input-group-append">
+                                        <button class="btn btn-outline-secondary" type="button">
+                                            <i class='la la-map-marker'></i>
+                                        </button>
+                                    </div>
+                                </div>
+                                <span class="text-danger" asp-validation-for="Address"></span>
+                                <input hidden asp-for="Lng" />
+                                <input hidden asp-for="Lat" />
+                            </div>
+                        </div>
+                        <div data-custom-wizard-step style="display:none;">
+                            <h3>@Localizer["Enter details"]</h3>
+                            <div class="form-group">
+                                <label class="required" asp-for="Price"></label>
+                                <input asp-for="Price" class="form-control" placeholder="@Localizer["Enter the price"]" />
+                                <span class="form-text text-danger" asp-validation-for="Price"></span>
+                            </div>
+                            <div class="form-group">
+                                <label asp-for="Description"></label>
+                                <textarea asp-for="Description" class="form-control"></textarea>
+                                <span class="form-text text-danger" asp-validation-for="Description"></span>
+                            </div>
+                            <!--end::Input-->
+                        </div>
+                        <div data-custom-wizard-step style="display:none;">
+                            <h3>@Localizer["Select amenities"]</h3>
+                            <div class="row" id="amenitiesContainer">
+
+                            </div>
+                            <!--end::Input-->
+                        </div>
+                        <div data-custom-wizard-step style="display:none;">
+                            <h3>@Localizer["Select requirements"]</h3>
+                            <span class="text-danger" asp-validation-for="RequirementIds"></span>
+                            <div class="row" id="requirementsContainer">
+
+                            </div>
+                            <!--end::Input-->
+                        </div>
+                        <div data-custom-wizard-step style="display:none;">
+                            <h3>@Localizer["Import medias"]</h3>
+                            <div class="row" id="mediasContainer">
+                                <div class="col-lg-4 col-md-6 filePreview mt-4">
+                                    <div class="card card-custom overlay">
+                                        <div class="card-body p-0">
+                                            <div class="overlay-wrapper">
+                                                <img src="" class="h-300px w-100 rounded" style="object-fit:cover;" />
+                                                <input hidden name="MediaUrls" type="checkbox" checked />
+                                            </div>
+                                            <div class="overlay-layer">
+                                                <button type="button" class="btn font-weight-bold btn-light-primary btn-shadow ml-2" data-delete-uploaded-file>
+                                                    <i class="flaticon-delete"></i>@Localizer["Delete"]
+                                                </button>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                            <!--end::Input-->
+                            <div class="form-group mt-5">
+                                <label class="btn btn-light-primary" id="formFileMultiple">
+                                    <input type="file" style="display:none;" id="mediafileinput" multiple>
+                                    <i class="la la-plus"></i>@Localizer["Add media files"]
+                                </label>
+                            </div>
+                        </div>
+                        <div data-custom-wizard-step style="display:none;">
+                            <div class="form-group">
+                                <label class="required" asp-for="ContactName"></label>
+                                <input asp-for="ContactName" class="form-control" />
+                                <span class="form-text text-danger" asp-validation-for="ContactName"></span>
+                            </div>
+                            <!--end::Input-->
+                            <!--begin::Input-->
+                            <div class="form-group">
+                                <label class="required" asp-for="ContactPhone"></label>
+                                <input id="phone" class="form-control" />
+                                <input hidden asp-for="ContactPhone" />
+                                <span class="form-text text-danger" asp-validation-for="ContactPhone"></span>
+                            </div>
+                            <!--end::Input-->
+                            <!--begin::Input-->
+                            <div class="form-group">
+                                <label class="required" asp-for="ContactEmail"></label>
+                                <input asp-for="ContactEmail" class="form-control" />
+                                <span class="form-text text-danger" asp-validation-for="ContactEmail"></span>
+                            </div>
+                            <!--end::Input-->
+                        </div>
+                        <div class="d-flex justify-content-between mt-5 pt-10">
+                            <div class="mr-2">
+                                <button type="button" class="btn btn-light-primary font-weight-bolder text-uppercase px-9 py-4" data-custom-wizard-previous>Previous</button>
+                            </div>
+                            <div>
+                                <button type="submit" class="btn btn-success font-weight-bolder text-uppercase px-9 py-4" data-custom-wizard-submit>Submit</button>
+                                <button type="button" class="btn btn-primary font-weight-bolder text-uppercase px-9 py-4" data-custom-wizard-next>Next</button>
+                            </div>
                         </div>
                     </div>
-                    <div class="uppy" id="kt_uppy_3">
-                        <div class="uppy-drag"></div>
-                        <div class="uppy-informer"></div>
-                        <div class="uppy-progress"></div>
-                        <div class="uppy-thumbnails"></div>
-                    </div>
-                    <div id="medias">
-                    </div>
+                </form>
+            </div>
 
-                </div>
-                <!--end: Wizard Step 5-->
-                <input asp-for="Lng" id="longtitudeInput" type="hidden" />
-                <input asp-for="Lat" id="latitudeInput" type="hidden" />
-                <!--begin: Wizard Actions-->
-                <div class="d-flex justify-content-between border-top mt-5 pt-10">
-                    <div class="mr-2">
-                        <button type="button" class="btn btn-light-primary font-weight-bolder text-uppercase px-9 py-4" data-wizard-type="action-prev">Previous</button>
-                    </div>
-                    <div>
-                        <button type="submit" class="btn btn-success font-weight-bolder text-uppercase px-9 py-4">Submit</button>
-                        <button type="button" class="btn btn-primary font-weight-bolder text-uppercase px-9 py-4" data-wizard-type="action-next">Next</button>
-                    </div>
-                </div>
-                <!--end: Wizard Actions-->
-            </form>
-            <!--end: Wizard Form-->
         </div>
     </div>
-    <!--end: Wizard Body-->
 </div>
