@@ -12,12 +12,12 @@
 
 @section Scripts{
     <script>//$.ajax({
-//            url: "https://www.googleapis.com/geolocation/v1/geolocate?key=AIzaSyCdVSHqtbLoxPRa5o35dyisKmSfHHgM0UI",
-//            type: 'POST',
-//            success: function (loc) {
-//                console.log(loc.location);
-//            },
-//        });</script>
+                                    //            url: "https://www.googleapis.com/geolocation/v1/geolocate?key=AIzaSyCdVSHqtbLoxPRa5o35dyisKmSfHHgM0UI",
+                                    //            type: 'POST',
+                                    //            success: function (loc) {
+                                    //                console.log(loc.location);
+                                    //            },
+                                    //        });</script>
 }
 
 <div class="bgi-size-cover bgi-position-center bg-image" style="background-image: url('/assets/media/static/homeintro_old.jpg');">
@@ -28,19 +28,17 @@
                     <!--begin::Heading Slogan-->
                     <h1 class="text-white font-weight-bolder mb-12 display-3">@tenant.Slogan</h1>
                     <!--end::Heading-->
-                    <a asp-action="Search" asp-controller="Rooms" class="btn btn-primary font-weight-bolder h4 text-uppercase">
-                        @Localizer["Search"]
+                    <a asp-action="Create" asp-controller="Rooms" class="btn btn-primary font-weight-bolder h4 text-uppercase">
                         <span class="svg-icon svg-icon-xl">
-                            <!--begin::Svg Icon | path:/metronic/theme/html/demo5/dist/assets/media/svg/icons/General/Search.svg-->
-                            <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px" height="24px" viewBox="0 0 24 24" version="1.1">
+                            <!--begin::Svg Icon | path:/var/www/preview.keenthemes.com/metronic/releases/2021-05-14-112058/theme/html/demo5/dist/../src/media/svg/icons/Navigation/Plus.svg--><svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px" height="24px" viewBox="0 0 24 24" version="1.1">
                                 <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-                                    <rect x="0" y="0" width="24" height="24"></rect>
-                                    <path d="M14.2928932,16.7071068 C13.9023689,16.3165825 13.9023689,15.6834175 14.2928932,15.2928932 C14.6834175,14.9023689 15.3165825,14.9023689 15.7071068,15.2928932 L19.7071068,19.2928932 C20.0976311,19.6834175 20.0976311,20.3165825 19.7071068,20.7071068 C19.3165825,21.0976311 18.6834175,21.0976311 18.2928932,20.7071068 L14.2928932,16.7071068 Z" fill="#000000" fill-rule="nonzero" opacity="0.3"></path>
-                                    <path d="M11,16 C13.7614237,16 16,13.7614237 16,11 C16,8.23857625 13.7614237,6 11,6 C8.23857625,6 6,8.23857625 6,11 C6,13.7614237 8.23857625,16 11,16 Z M11,18 C7.13400675,18 4,14.8659932 4,11 C4,7.13400675 7.13400675,4 11,4 C14.8659932,4 18,7.13400675 18,11 C18,14.8659932 14.8659932,18 11,18 Z" fill="#000000" fill-rule="nonzero"></path>
+                                    <rect fill="#000000" x="4" y="11" width="16" height="2" rx="1" />
+                                    <rect fill="#000000" opacity="0.3" transform="translate(12.000000, 12.000000) rotate(-270.000000) translate(-12.000000, -12.000000) " x="4" y="11" width="16" height="2" rx="1" />
                                 </g>
                             </svg>
                             <!--end::Svg Icon-->
                         </span>
+                        @Localizer["Create Announcement"]
                     </a>
                 </div>
 
@@ -79,25 +77,10 @@
                 <h1 class="font-weight-bolder">@Localizer["LastAnnouncements"]</h1>
             </div>
             <div class="row">
-                @for (int i = 0; i < 9; i++)
-                {
-                    Model.Rooms.Add(new Application.CQRS.Rooms.Queries.GetActiveRooms.RoomBriefVM
-                    {
-                        Id = Guid.NewGuid().ToString(),
-                        MediaUrl = "/assets/media/static/homeintro_old.jpg",
-                        MediaAltTag = "asda",
-                        UpdatedAt = DateTime.Now,
-                        Price = 123,
-                        Title = "Oglan otaq yoldasi axtarilir",
-                        Url = "#"
-                    });
-                }
                 @foreach (var room in Model.Rooms)
                 {
                     <div class="col-md-4 col-lg-12 col-xxl-4">
-
                         <div class="card card-custom gutter-b card-stretch">
-
                             <div class="card-body d-flex flex-column rounded bg-light justify-content-between  p-2">
                                 <div class="text-center rounded mb-7">
                                     <img loading="lazy" src="@room.MediaUrl" alt="@room.MediaAltTag" class="w-100 h-300px img-fluid rounded" style="object-fit:cover;">
@@ -107,7 +90,9 @@
                                         <h4 class="font-size-h5">
                                             <a href="@room.Url" class="text-dark-75 font-weight-bolder">@room.Title</a>
                                         </h4>
-                                        <div class="font-size-h6 text-muted font-weight-bolder">@room.Price</div>
+                                        <div class="font-size-h6 text-muted font-weight-bolder">
+                                            @room.Price AZN
+                                        </div>
                                     </div>
                                     <span class="d-flex justify-content-end mb-2"><i class="la la-bookmark-o icon-3x"></i></span>
                                 </div>
@@ -117,7 +102,7 @@
                 }
             </div>
             <div class="d-flex justify-content-end">
-                <a class="btn btn-link pr-0"><i>@Localizer["ShowAll"]</i><i class="pr-0 pl-2 icon-md la la-angle-down"></i> </a>
+                <a class="btn btn-link pr-0" asp-action="Search" asp-controller="Rooms"><i>@Localizer["Show All"]</i><i class="pr-0 pl-2 icon-md la la-angle-down"></i> </a>
             </div>
         </div>
     </div>
