@@ -61,7 +61,7 @@
         <!--begin::Nav-->
         <div class="navi navi-spacer-x-0 p-0">
             <!--begin::Item-->
-            <a href="custom/apps/user/profile-1/personal-information.html" class="navi-item">
+            <a asp-action="PersonalInfo" asp-controller="MyProfile" asp-area="" class="navi-item">
                 <div class="navi-link">
                     <div class="symbol symbol-40 bg-light mr-3">
                         <div class="symbol-label">
@@ -79,9 +79,9 @@
                         </div>
                     </div>
                     <div class="navi-text">
-                        <div class="font-weight-bold">@Localizer["MyProfile"]</div>
+                        <div class="font-weight-bold">@Localizer["My Profile"]</div>
                         <div class="text-muted">
-                            @Localizer["AccountSettingsAndMore"]
+                            @Localizer["Account settings and more"]
                         </div>
                     </div>
                 </div>
